Cellbow python source code
This is the python sorece code of Cellbow. User can use this file to process large dataset. And training their own Cellbow. Compared with the website or Imagej plugin version, the python version has the advantage of faster processing speed, both prediction and training. The dis advantage is user need to install python and tensorflow before use it. It the user get satisfactory results from the website or imagej plugin, we recommend the user to use this python code for large dataset image processing.